put implementation of abt and hg config in .c files
