+++ src/Makefile.subdir
@@ -17,6 +17,8 @@ noinst_HEADERS += src/margo-abt-macros.h \
 src_libmargo_la_SOURCES += \
  src/margo-core.c \
  src/margo-config.c \
+ src/margo-abt-config.c \
+ src/margo-hg-config.c \
  src/margo-abt-profiling.c \
  src/margo-bulk-pool.c \
  src/margo-globals.c \
